Fix compiler warnings

+++ include/isect2d.h
@@ -21,8 +21,8 @@ inline static std::set<std::pair<int, int>> intersect(const std::vector<isect2d:
     }
 
 #ifdef BRUTEFORCE
-    for (int i = 0; i < _aabbs.size(); ++i) {
-        for (int j = i + 1; j < _aabbs.size(); ++j) {
+    for (size_t i = 0; i < _aabbs.size(); ++i) {
+        for (size_t j = i + 1; j < _aabbs.size(); ++j) {
             if (_aabbs[i].intersect(_aabbs[j])) {
                 pairs.insert({ i, j });
             }
@@ -38,7 +38,7 @@ inline static std::set<std::pair<int, int>> intersect(const std::vector<isect2d:
     isect2d::BVHNode* node;
     std::stack<isect2d::BVHNode*> todo;
 
-    for (int i = 0; i < _aabbs.size(); ++i) {
+    for (size_t i = 0; i < _aabbs.size(); ++i) {
         const isect2d::AABB& aabb = _aabbs[i];
         todo.push(bvh.getRoot());
